naat/pyproject.toml
# https://python-poetry.org/docs/pyproject/
[tool.poetry]
name = "naat"
version = "0.1.0"
description = "Sandbox"
authors = ["Clément Brutti-Mairesse <clement.brutti.mairesse@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ClementBM/sandbox"
repository = "https://github.com/ClementBM/sandbox"
keywords = ["mykeyword", "mykeyword", "mykeyword"]
include = [
    "LICENSE",
]
packages = [
    { include = "naat_core" },
    { include = "naat_data" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
dstb = { git = "https://ClementBM@github.com/ClementBM/datascience-toolbox", branch = "main" }
python-docx = "^0.8.11"
pypdf = "^3.4.1"
validator-collection = "^1.5.0"
python-dotenv = "^0.21.1"
pip = "^23.0"
rake-nltk = "^1.0.6"
gcld3 = "^3.0.13"
pytesseract = "^0.3.10"
pdf2image = "^1.16.2"
transformers = "^4.26.1"
torch = "^1.13.1"
google-api-python-client = "^2.80.0"
pdfminer-six = "^20221105"
pdf2docx = "^0.5.6"
tika = "^2.6.0"

[tool.poetry.dev-dependencies]
pylint = "*"
pytest = "*"
mypy = "*"
jupyter = "*"
notebook = "*"
ipykernel = "*"
black = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"